+++ src/main/java/com/unb/tracker/model/Course.java
@@ -1,4 +1,4 @@
-package com.unb.tracker;
+package com.unb.tracker.model;
 
 import javax.persistence.*;
 import java.util.List;
