+++ src/main/java/com/unb/tracker/model/Seat.java
@@ -1,4 +1,4 @@
-package com.unb.tracker;
+package com.unb.tracker.model;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
