+++ src/main/java/com/unb/tracker/repository/CourseRepository.java
@@ -1,6 +1,7 @@
-package com.unb.tracker;
+package com.unb.tracker.repository;
 
 
+import com.unb.tracker.model.Course;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
