+++ src/main/java/com/unb/tracker/repository/UserRepository.java
@@ -1,10 +1,10 @@
-package com.unb.tracker;
+package com.unb.tracker.repository;
 
 
+import com.unb.tracker.model.User;
 import org.springframework.data.repository.CrudRepository;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 // CRUD refers Create, Read, Update, Delete
 
 public interface UserRepository extends CrudRepository<User, Long> {
-
 }
